modify scenariogateway w.r.t object boundingbox

+++ EnvironmentSimulator/ScenarioEngineDLL/scenarioenginedll.hpp
@@ -39,12 +39,12 @@ typedef struct
 	float laneOffset;
 	float s;
 	float speed;
-	float centerOffsetX;
-	float centerOffsetY;
-	float centerOffsetZ;
-	float width;
-	float length;
-	float height;
+	double centerOffsetX;
+	double centerOffsetY;
+	double centerOffsetZ;
+	double width;
+	double length;
+	double height;
 } SE_ScenarioObjectState;
 
 
@@ -132,7 +132,7 @@ extern "C"
 	*/
 	SE_DLL_API const char* SE_GetOSIRoadLane(int* size, int object_id);
 	/**
-	The SE_GetOSIRoadLane function returns a char array containing the osi Lane Boundary information/message with the specified GLOBAL id 
+	The SE_GetOSIRoadLane function returns a char array containing the osi Lane Boundary information/message with the specified GLOBAL id
 	*/
 	SE_DLL_API const char* SE_GetOSILaneBoundary(int* size, int global_id);
 	/**
